搜索
搜索接收处理封装分页数据返回到前台

+++ dg-common/src/main/java/com/vivi/common/vo/PageResult.java
@@ -7,7 +7,7 @@ import java.util.List;
 @Data
 public class PageResult<T> {
     private Long total;// 总条数
-    private Long totalPage;// 总页数
+    private int totalPage;// 总页数
     private List<T> items;// 当前页数据
 
     public PageResult() {
@@ -18,7 +18,7 @@ public class PageResult<T> {
         this.items = items;
     }
 
-    public PageResult(Long total, Long totalPage, List<T> items) {
+    public PageResult(Long total, int totalPage, List<T> items) {
         this.total = total;
         this.totalPage = totalPage;
         this.items = items;
@@ -40,11 +40,11 @@ public class PageResult<T> {
         this.items = items;
     }
 
-    public Long getTotalPage() {
+    public int getTotalPage() {
         return totalPage;
     }
 
-    public void setTotalPage(Long totalPage) {
+    public void setTotalPage(int totalPage) {
         this.totalPage = totalPage;
     }
 }
